+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.0.3 - 1.21
+### Additions
+- Added built-in support for No Man's Land mobs and biomes
+- Added new exclusive textures for No Man's Land fox variants!
+
 # 1.0.2 - 1.21
 ### Additions
 - Added support for No Man's Land fox variants!